pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teamautobot"
version = "0.1.0"
description = "TeamAutobot runtime and CLI foundation slice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "azure-identity>=1.21.0",
  "openai>=1.99.1",
  "python-dotenv>=1.0.0",
]

[project.scripts]
teamautobot = "teamautobot.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.uv]
package = true
